Updated for 64bit processor.

+++ CTAssetsPickerController/NSDate+TimeInterval.m
@@ -53,12 +53,12 @@
     
     if (components.hour > 0)
     {
-        return [NSString stringWithFormat:@"%d:%02d:%02d", components.hour, components.minute, components.second];
+        return [NSString stringWithFormat:@"%ld:%02ld:%02ld", (long)components.hour, (long)components.minute, (long)components.second];
     }
     
     else
     {
-        return [NSString stringWithFormat:@"%d:%02d", components.minute, components.second];
+        return [NSString stringWithFormat:@"%ld:%02ld", (long)components.minute, (long)components.second];
     }    
 }
